Changes for user add in db

+++ src/main/sql/Users.sql
@@ -1,5 +1,5 @@
 CREATE TABLE `users` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
+  `id` bigint NOT NULL AUTO_INCREMENT,
   `user` varchar(30) NOT NULL,
   `email` varchar(30) DEFAULT NULL,
   `password` varchar(100) NOT NULL,
